Day 17
Write a Calculator class with a single method: int power(int,int). The power method takes two integers,  and , as parameters and returns the integer result of . If either n or p  is negative, then the method must throw an exception with the message: n and p should be non-negative.